Lab 1
It seemed silly that the user couldn’t input grades, so I put in
optional code to enter grades on a loop. In the future I’ll exchange
this for code that can read an ascii fill that would also include names
of students. I used the weights we discussed in class as now weights
were given in the assignment. The histogram saves, but I can’t see an
image when I open the file

In ex3a I also did optional data inputs for fun

In ex3b I had similar issues with the savefig command line. I can’t see
the image. The ascii file does save and is openable, but not very
organized.